meta: Corrected frame delays slightly.

+++ libplip/Cpu/PlipCpu.cpp
@@ -22,7 +22,7 @@ void PlipCpu::Step() {
     }
 }
 
-long PlipCpu::GetCycleTime() const {
+double PlipCpu::GetCycleTime() const {
     return m_cycle;
 }
 
@@ -32,5 +32,5 @@ long PlipCpu::GetHz() const {
 
 void PlipCpu::SetHz(const long hz) {
     m_hz = hz;
-    m_cycle = 1000000000 / hz;
+    m_cycle = 1000000000.0 / hz;
 }
